% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{CACE}
\alias{CACE}
\title{Wrapper}
\usage{
CACE(y, a, z, cov, delta = 2)
}
\arguments{
\item{y}{your outcome variable (a vector)}

\item{a}{your treatment variable (a vector)}

\item{z}{your instrument (a vector)}

\item{cov}{a dataframe of covariates}

\item{delta}{is the amount you want to shift by}
}
\value{
an estimate of the causal effect
}
\description{
Wraps up the whole process of estimating phi
using sample splitting.
}

