% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_est.R
\name{mean_est}
\alias{mean_est}
\title{Regression Function Estimates}
\usage{
mean_est(y, a, z, cov)
}
\arguments{
\item{y}{the outcome variable in your training set}

\item{a}{the treatment variable in your training set}

\item{z}{the instrument in your training set}

\item{cov}{the covariates in your training set. Should be a
data frame and it should not include any factors for now.}
}
\value{
gives you ymean and amean in the global envr. Note that I
tried to return these as a list and it gave me something weird, but
this works.
}
\description{
Use a SuperLearner to estimate the mean
functions for the outcome and the treatment given covariates
and the instrument
}

