% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{mysplit}
\alias{mysplit}
\title{Split data into 2}
\usage{
mysplit(dat)
}
\arguments{
\item{dat}{your dataset}
}
\value{
creates two datasets in the global environment called ds1 and ds2
}
\description{
Splits your dataset into 2 randomly
}

