% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_preds.R, R/phi_est.R
\name{phi_est}
\alias{phi_est}
\title{Estimate Phi on half of data}
\usage{
phi_est(y, a, z, cov, ymean, amean, p, delta = 20)

phi_est(y, a, z, cov, ymean, amean, p, delta = 20)
}
\arguments{
\item{y}{is the outcome vector}

\item{a}{is the treatment vector}

\item{z}{is the instrument vector}

\item{cov}{is a dataframe of covariates}

\item{ymean}{is the mean estimates for the outcome
variable from the mean_est function}

\item{amean}{is the mean estimates for the treatment
variable from the mean_est function}

\item{p}{the conditional density estimation object from the
propscore_est funciton}

\item{delta}{the level you want to shift by}

\item{dat}{is a dataset that is being used as your training
dataset overall}

\item{m}{a list of your mean estimates from the mean_est function}

\item{p}{the conditional density estimation object from the
propscore_est funciton}
}
\value{
an estimate of the causal effect

an estimate of the causal effect
}
\description{
Take in the estimates done of the nuisance function on
the training data. Get predictions using these on the test data.
Input those into the formula for your real parameter. Output the
estimate of your parameter for this train/test combo.

Take in the estimates done of the nuisance function on
the training data. Get predictions using these on the test data.
Input those into the formula for your real parameter. Output the
estimate of your parameter for this train/test combo.
}

