% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propscore_est.R
\name{propscore_est}
\alias{propscore_est}
\title{Conditional Density Estimates}
\usage{
propscore_est(y, x, quiet = T)
}
\arguments{
\item{y}{is your outcome variable (here it should be your instrument)}

\item{x}{is a dataframe of covariates}
}
\value{
conditional density object
}
\description{
Use FlexCoDE to estimate the conditional densities
that give the propensity scores
}

