% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{split}
\alias{split}
\title{Split data}
\usage{
split(dat)
}
\arguments{
\item{k}{either 1 or 2, the half of the dataset you want to work with}
}
\value{
returns 2 halves of your dataset as a list
}
\description{
Splits your dataset into 2 and attaches the kth one
}

