% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapper}
\alias{wrapper}
\title{Wrapper}
\usage{
wrapper(dat)
}
\value{
an estimate of the causal effect
}
\description{
Wraps up the whole process of estimating phi
using sample splitting. For now the covariates are hard
coded, as are a bunch of things. Will fix later. I'm taking
out the loading step so that other people can use this, but
it still has to be used on the visits dataset
}

