% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{mysplit}
\alias{mysplit}
\title{Split data}
\usage{
mysplit(dat)
}
\arguments{
\item{dat}{your dataset}
}
\value{
returns 2 halves of your dataset as a list, attaches split1
}
\description{
Splits your dataset into 2 and attaches the kth one
}

